requirements.txt
streamlit
python-dotenv
youtube-transcript-api
reportlab
google-generativeai
exa-py
